Issue #1356: Bugfix

+++ casadi/solvers/newton.cpp
@@ -146,8 +146,8 @@ namespace casadi {
       }
 
       // Factorize the linear solver with J
-      fill_n(arg1, LINSOL_NUM_IN, nullptr);
-      fill_n(res1, LINSOL_NUM_OUT, nullptr);
+      fill_n(arg1, static_cast<int>(LINSOL_NUM_IN), nullptr);
+      fill_n(res1, static_cast<int>(LINSOL_NUM_OUT), nullptr);
       arg1[LINSOL_A] = jac;
       linsol_.linsol_prepare(0, arg1, res1, iw, w);
       linsol_.linsol_solve(f, 1, false);
